+++ cats/optimise_starttime.py
